+++ src/test/java/ZUV/MiscValidatorTest.java
@@ -12,9 +12,13 @@ public class MiscValidatorTest extends ResourceCase  {
 
 		ZUGFeRDValidator zfv=new ZUGFeRDValidator();
 
-		String res=zfv.validate("/dhfkbv/sfjkh");
-		assertEquals("<validation><messages><error type=\"1\">File /dhfkbv/sfjkh not found</error>\n" + 
-				"</messages><summary status='invalid'/></validation>", res);
+		String res=zfv.validate(null);
+		assertTrue(res.matches("<validation filename='' datetime='.*?'><messages><error type=\"10\">Filename not specified</error>\n" + 
+				"</messages><summary status='invalid'/></validation>"));
+
+		res=zfv.validate("/dhfkbv/sfjkh");
+		assertTrue(res.matches("<validation filename='sfjkh' datetime='.*?'><messages><error type=\"1\">File not found</error>\n" + 
+								"</messages><summary status='invalid'/></validation>"));
 
 		File tempFile=null;
 		try {
@@ -25,9 +29,9 @@ public class MiscValidatorTest extends ResourceCase  {
 		}
 
 		res=zfv.validate(tempFile.getAbsolutePath());
-		assertEquals("<validation><messages>" + 
-				"<error type=\"5\">File "+tempFile.getAbsolutePath()+" too small</error>\n" + 
-				"</messages><summary status='invalid'/></validation>", res);
+		assertTrue(res.matches("<validation.*?><messages>" + 
+				"<error type=\"5\">File too small</error>\n" + 
+				"</messages><summary status='invalid'/></validation>"));
 		
 		
 		String fileContent = "ladhvkdbfk  wkhfbkhdhkb svbkfsvbksfbvk sdvsdvbksjdvbkfdsv sdvbskdvbsjhkvbfskh dvbskfvbkfsbvke"
@@ -45,8 +49,8 @@ public class MiscValidatorTest extends ResourceCase  {
 		 
 
 		res=zfv.validate(tempFile.getAbsolutePath());
-		assertEquals("<validation><messages><exception type=\"8\">File does not look like PDF nor XML (contains neither %PDF nor <?xml)</exception>\n" + 
-				"</messages><summary status='invalid'/></validation>", res);
+		assertTrue(res.matches("<validation filename='"+tempFile.getName()+"' .*?><messages><exception type=\"8\">File does not look like PDF nor XML .*</exception>\n" + 
+				"</messages><summary status='invalid'/></validation>"));
 
 		
 		// clean up